src/async_api/requirements.txt
aiohttp==3.7.2
aioredis==1.3.1
backoff==1.10.0
elasticsearch-async==6.2.0
elasticsearch==7.12.0
fastapi-pagination==0.7.0
fastapi==0.63.0
orjson==3.3.1
pydantic==1.8.1
pytest-asyncio==0.12.0
pytest==6.1.2
repackage==0.7.3
uvicorn==0.13.4
grpc-interceptor~= 0.12.0
grpcio-tools~= 1.30